+++ cgnet/tests/test_feature_utils.py
@@ -33,7 +33,7 @@ def test_radial_basis_function():
     # Make sure radial basis functions are consistent with manual calculation
 
     # Distances need to have shape (n_batch, n_beads, n_neighbors)
-    distances = torch.randn((frames, beads, beads - 1))
+    distances = torch.randn((frames, beads, beads - 1), dtype=torch.float64)
     # Define random parameters for the RBF
     variance = np.random.random() + 1
     n_gaussians = np.random.randint(5, 10)
@@ -50,7 +50,7 @@ def test_radial_basis_function():
     # e_k (r_j - r_i) = exp(- \gamma (\left \| r_j - r_i \right \| - \mu_k)^2)
     # with centers mu_k calculated on a uniform grid between
     # zero and the distance cutoff and gamma as a scaling parameter.
-    centers = np.linspace(low_cutoff, high_cutoff, n_gaussians)
+    centers = np.linspace(low_cutoff, high_cutoff, n_gaussians).astype(np.float64)
     gamma = -0.5 / variance
     distances = np.expand_dims(distances, axis=3)
     magnitude_squared = (distances - centers)**2
@@ -66,7 +66,7 @@ def test_radial_basis_function_distance_masking():
     # expanded distances returned by GaussianRBF are zero
 
     # Distances need to have shape (n_batch, n_beads, n_neighbors)
-    distances = torch.randn((frames, beads, beads - 1))
+    distances = torch.randn((frames, beads, beads - 1), dtype=torch.float64)
     # Define random parameters for the RBF
     variance = np.random.random() + 1
     high_cutoff = np.random.uniform(5.0, 10.0)
@@ -99,7 +99,7 @@ def test_radial_basis_function_normalize():
     # normalized if 'normalize_output' is specified as True
 
     # Distances need to have shape (n_batch, n_beads, n_neighbors)
-    distances = torch.randn((frames, beads, beads - 1))
+    distances = torch.randn((frames, beads, beads - 1), dtype=torch.float64)
     # Define random parameters for the RBF
     variance = np.random.random() + 1
     n_gaussians = np.random.randint(5, 10)
@@ -117,7 +117,7 @@ def test_radial_basis_function_normalize():
     # e_k (r_j - r_i) = exp(- \gamma (\left \| r_j - r_i \right \| - \mu_k)^2)
     # with centers mu_k calculated on a uniform grid between
     # zero and the distance cutoff and gamma as a scaling parameter.
-    centers = np.linspace(low_cutoff, high_cutoff, n_gaussians)
+    centers = np.linspace(low_cutoff, high_cutoff, n_gaussians).astype(np.float64)
     gamma = -0.5 / variance
     distances = np.expand_dims(distances, axis=3)
     magnitude_squared = (distances - centers)**2
@@ -135,7 +135,7 @@ def test_polynomial_cutoff_rbf():
     # manual calculations
 
     # Distances need to have shape (n_batch, n_beads, n_neighbors)
-    distances = np.random.randn(frames, beads, beads - 1).astype('float32')
+    distances = np.random.randn(frames, beads, beads - 1).astype(np.float64)
     # Define random parameters for the polynomial_cutoff RBF
     n_gaussians = np.random.randint(5, 10)
     high_cutoff = np.random.uniform(5.0, 10.0)
@@ -153,7 +153,7 @@ def test_polynomial_cutoff_rbf():
     # Manually calculate expansion with numpy
     # First, we compute the centers and the scaling factors
     centers = np.linspace(np.exp(-high_cutoff), np.exp(-low_cutoff),
-                          n_gaussians)
+                          n_gaussians).astype(np.float64)
     beta = np.power(((2/n_gaussians) * (1-np.exp(-high_cutoff))), -2)
 
     # Next, we compute the gaussian portion
@@ -191,7 +191,7 @@ def test_polynomial_cutoff_rbf_distance_masking():
     # expanded distances returned by PolynomialCutoffRBF are zero
 
     # Distances need to have shape (n_batch, n_beads, n_neighbors)
-    distances = torch.randn((frames, beads, beads - 1))
+    distances = torch.randn((frames, beads, beads - 1), dtype=torch.float64)
     # Define random parameters for the RBF
     n_gaussians = np.random.randint(5, 10)
     high_cutoff = np.random.uniform(5.0, 10.0)
@@ -215,7 +215,7 @@ def test_polynomial_cutoff_rbf_distance_masking():
     # Manually calculate expansion with numpy
     # First, we compute the centers and the scaling factors
     centers = np.linspace(np.exp(-high_cutoff), np.exp(-low_cutoff),
-                          n_gaussians)
+                          n_gaussians).astype(np.float64)
     beta = np.power(((2/n_gaussians) * (1-np.exp(-high_cutoff))), -2)
 
     # Next, we compute the gaussian portion
@@ -251,7 +251,7 @@ def test_polynomial_cutoff_rbf_normalize():
     # normalized if 'normalize_output' is specified as True
 
     # Distances need to have shape (n_batch, n_beads, n_neighbors)
-    distances = np.random.randn(frames, beads, beads - 1).astype('float32')
+    distances = np.random.randn(frames, beads, beads - 1).astype(np.float64)
     # Define random parameters for the polynomial_cutoff RBF
     n_gaussians = np.random.randint(5, 10)
     high_cutoff = np.random.uniform(5.0, 10.0)
@@ -270,7 +270,7 @@ def test_polynomial_cutoff_rbf_normalize():
     # Manually calculate expansion with numpy
     # First, we compute the centers and the scaling factors
     centers = np.linspace(np.exp(-high_cutoff), np.exp(-low_cutoff),
-                          n_gaussians)
+                          n_gaussians).astype(np.float64)
     beta = np.power(((2/n_gaussians) * (1-np.exp(-high_cutoff))), -2)
 
     # Next, we compute the gaussian portion
@@ -307,7 +307,6 @@ def test_polynomial_cutoff_rbf_normalize():
                                polynomial_cutoff_rbf_manual, rtol=1e-5)
 
 
-
 def test_polynomial_cutoff_rbf_zero_cutoff():
     # This test ensures that a choice of zero cutoff produces
     # a set of basis functions that all occupy the same center
@@ -322,7 +321,7 @@ def test_polynomial_cutoff_rbf_zero_cutoff():
     np.testing.assert_equal(np.inf, polynomial_cutoff_rbf.beta)
 
     # Next we make a mock array of centers at 1.0
-    centers = torch.linspace(np.exp(-cutoff), np.exp(-cutoff), n_gaussians)
+    centers = torch.linspace(np.exp(-cutoff), np.exp(-cutoff), n_gaussians, dtype=torch.float64)
 
     # Here, we test to see that centers are equal in this corner case
     np.testing.assert_equal(centers.numpy(),
@@ -334,7 +333,7 @@ def test_shifted_softplus():
     # manual calculation
 
     # Initialize random feature vector
-    feature = torch.randn((frames, beads), dtype=torch.double)
+    feature = torch.randn((frames, beads), dtype=torch.float64)
 
     ssplus = ShiftedSoftplus()
     # Shifted softplus has the following form:
